# Various `JSON.generate` optimizations